refactor: route seperation from server.js
